Leveling system/Rep system/Stats system
This push includes the leveling system, rep system and stat system. It
also includes the reworked chat which has to be checked.
The system was based on an earlier build of raindrop, and as such we
have to make sure all the dependencies are correctly added to the newer
builds.